# dgc-extracted/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DGC Arcade API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 24
     */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  balance: number;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  totalBets?: number | null;
  /** @nullable */
  totalWon?: number | null;
  role: string;
  isBanned: boolean;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface Game {
  id: number;
  slug: string;
  name: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  minBet: number;
  maxBet: number;
  houseEdge?: number;
  active: boolean;
}

/**
 * Game-specific options (e.g. chosen side for coin flip)
 * @nullable
 */
export type BetInputMeta = {[key: string]: unknown} | null;

export interface BetInput {
  gameId: number;
  amount: number;
  /** @nullable */
  clientSeed?: string | null;
  /**
     * Game-specific options (e.g. chosen side for coin flip)
     * @nullable
     */
  meta?: BetInputMeta;
}

/**
 * @nullable
 */
export type BetRecordMeta = {[key: string]: unknown} | null;

export interface BetRecord {
  id: number;
  userId: number;
  username: string;
  gameId: number;
  gameName: string;
  amount: number;
  payout: number;
  won: boolean;
  /** @nullable */
  multiplier?: number | null;
  /** @nullable */
  serverSeed?: string | null;
  /** @nullable */
  clientSeed?: string | null;
  /** @nullable */
  meta?: BetRecordMeta;
  createdAt: string;
}

export interface BetResult {
  bet: BetRecord;
  newBalance: number;
  won: boolean;
  payout: number;
  multiplier: number;
}

export interface DepositInput {
  amount: number;
  /** e.g. BTC, ETH, LTC, USDT */
  currency: string;
}

export interface DepositLink {
  paymentUrl: string;
  trackId: string;
}

export interface OxapayCallbackInput {
  /** @nullable */
  trackId?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  orderId?: string | null;
}

export interface WithdrawInput {
  amount: number;
  currency: string;
  address: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  bet_win: 'bet_win',
  bet_loss: 'bet_loss',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  userId: number;
  type: TransactionType;
  amount: number;
  currency: string;
  status: TransactionStatus;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  totalWon: number;
  totalBets: number;
}

export interface PlatformStats {
  totalPlayers: number;
  totalBets: number;
  totalWagered: number;
  biggestWin: number;
}

export type ListBetsParams = {
limit?: number;
};

export type ListRecentBetsAllParams = {
limit?: number;
};

export type ListTransactionsParams = {
limit?: number;
};

export type GetLeaderboardParams = {
limit?: number;
};