# dgc-extracted/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DGC Arcade API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Games ─────────────────────────────────────────────────────────────
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List the authenticated user's transaction history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, username, balance, createdAt, role, isBanned]
      properties:
        id:
          type: integer
        username:
          type: string
        balance:
          type: number
        avatarUrl:
          type: ["string", "null"]
        totalBets:
          type: ["integer", "null"]
        totalWon:
          type: ["number", "null"]
        role:
          type: string
        isBanned:
          type: boolean
        createdAt:
          type: string

    # ── Games ─────────────────────────────────────────────────────────
...
    Transaction:
      type: object
      required: [id, userId, type, amount, currency, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, bet_win, bet_loss]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        txHash:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Leaderboard ───────────────────────────────────────────────────
...